+++ CHANGELOG.md
@@ -40,6 +40,10 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
   `github.com/signalfx/splunk-otel-go/instrumentation/github.com/syndtr/goleveldb/leveldb/splunkleveldb`
   instrumentation for the `github.com/syndtr/goleveldb/leveldb`
   package. (#186)
+- Add the
+  `github.com/signalfx/splunk-otel-go/instrumentation/github.com/tidwall/buntdb/splunkbuntdb`
+  instrumentation for the `github.com/tidwall/buntdb`
+  package. (#217)
 
 ### Changed